requirements.txt
APScheduler==3.10.1
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.52.0
django-apscheduler==0.6.2
django-bootstrap-modal-forms==2.2.0
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-crispy-forms==2.0
djangorestframework==3.14.0
fontawesomefree==6.3.0
html5lib==1.1
idna==3.4
lxml==4.9.2
oauthlib==3.2.2
oscrypto==1.3.0
Pillow==9.3.0
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.6.0
pypdf==3.8.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.28.2
requests-oauthlib==1.3.1
shortuuid==1.0.11
six==1.16.0
soupsieve==2.4
sqlparse==0.4.2
stripe==5.5.0
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2022.5
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
webencodings==0.5.1
xhtml2pdf==0.2.10
